+++ contracts/mocks/fake-tokens/Particle.sol
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: GPL3
 pragma solidity ^0.8.19;
 
-import "@openzeppelin/contracts/token/ERC721/ERC721.sol";
-import "@openzeppelin/contracts/access/Ownable.sol";
+import {ERC721} from "@openzeppelin/contracts/token/ERC721/ERC721.sol";
+import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
 
-import "../../ERC6551/IERC6551Account.sol";
+import {IERC6551Account} from "../../ERC6551/IERC6551Account.sol";
 
 contract Particle is ERC721, Ownable {
   string private _baseTokenURI;
@@ -22,7 +22,7 @@ contract Particle is ERC721, Ownable {
   }
 
   // testing a delegated call to transfer the NFT
-  function transferFromBoundAccount(address from, address to, uint tokenId) public {
+  function transferFromBoundAccount(address from, address to, uint256 tokenId) public {
     IERC6551Account(payable(from)).executeCall(
       address(this),
       0,
