+++ contracts/utils/TokenUtils.sol
@@ -3,16 +3,16 @@ pragma solidity ^0.8.19;
 
 // Author: Francesco Sullo <francesco@sullo.co>
 
-import "@openzeppelin/contracts/utils/introspection/IERC165.sol";
-import "@openzeppelin/contracts/token/ERC721/IERC721.sol";
-import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
-import "@openzeppelin/contracts/token/ERC1155/IERC1155.sol";
+import {IERC165} from "@openzeppelin/contracts/utils/introspection/IERC165.sol";
+import {IERC721} from "@openzeppelin/contracts/token/ERC721/IERC721.sol";
+import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
+import {IERC1155} from "@openzeppelin/contracts/token/ERC1155/IERC1155.sol";
 
-import "../protected-nft/IProtectedERC721.sol";
-import "../vaults/IFlexiVault.sol";
-import "../protected-nft/IActors.sol";
+import {IProtectedERC721} from "../protected-nft/IProtectedERC721.sol";
+import {IFlexiVault} from "../vaults/IFlexiVault.sol";
+import {IActors} from "../protected-nft/IActors.sol";
 
-//import "hardhat/console.sol";
+//import {console} from "hardhat/console.sol";
 
 library TokenUtils {
   error TheERC721IsAProtector();
@@ -62,7 +62,7 @@ library TokenUtils {
     uint256 amount,
     address beneficiary,
     uint256 timestamp,
-    uint validFor
+    uint256 validFor
   ) external view returns (bytes32) {
     return
       keccak256(
@@ -75,7 +75,7 @@ library TokenUtils {
     address recipient,
     IActors.Level level,
     uint256 timestamp,
-    uint validFor
+    uint256 validFor
   ) external view returns (bytes32) {
     return keccak256(abi.encode("\x19\x01", block.chainid, address(this), owner, recipient, level, timestamp, validFor));
   }
@@ -85,7 +85,7 @@ library TokenUtils {
     address beneficiary,
     IActors.Status status,
     uint256 timestamp,
-    uint validFor
+    uint256 validFor
   ) external view returns (bytes32) {
     return keccak256(abi.encode("\x19\x01", block.chainid, address(this), owner, beneficiary, status, timestamp, validFor));
   }
@@ -98,7 +98,7 @@ library TokenUtils {
     uint256[] memory amounts,
     address[] memory beneficiaries,
     uint256 timestamp,
-    uint validFor
+    uint256 validFor
   ) external view returns (bytes32) {
     return
       keccak256(
@@ -118,11 +118,11 @@ library TokenUtils {
       );
   }
 
-  function hashEjectRequest(uint256 owningTokenId, uint256 timestamp, uint validFor) external view returns (bytes32) {
+  function hashEjectRequest(uint256 owningTokenId, uint256 timestamp, uint256 validFor) external view returns (bytes32) {
     return keccak256(abi.encodePacked("\x19\x01", block.chainid, address(this), owningTokenId, timestamp, validFor));
   }
 
-  function hashTransferRequest(uint256 tokenId, address to, uint256 timestamp, uint validFor) public view returns (bytes32) {
+  function hashTransferRequest(uint256 tokenId, address to, uint256 timestamp, uint256 validFor) public view returns (bytes32) {
     return keccak256(abi.encode("\x19\x01", block.chainid, address(this), tokenId, to, timestamp, validFor));
   }
 }
